Replaced GP1 with GP2. Other fixes and checks
