+++ input/jgr_grid.cfg
@@ -1,0 +1,188 @@
+########## PROTEUS configuration file
+
+##### Star configuration
+
+star_model                  = 1                 # Evolution model to use for star (0: constant, 1: Mors, 2: Baraffe)
+star_rot_percentile         = 60.0              # Distribution percentile used to find rotation rate
+star_mass                   = 1                 # M_sun, mass of star
+star_radius_modern          = 1                 # R_sun, radius of star (TODAY)  
+star_temperature_modern     = 5772.0            # K, temperature of star (TODAY)
+star_luminosity_modern      = 1                 # L_sun, luminosity of star (TODAY) used for Baraffe model
+star_age_modern             = 4.567e9             # yr, age of star (TODAY)
+star_spectrum               = input/spec_sun.txt   # Stellar spectrum file at 1 AU (TODAY)
+star_btrack                 = input/BHAC15-M1p000000.txt
+
+
+##### Planet configuration
+mean_distance               = 1         # AU, star-planet distance
+mass                        = 5.972e24  # kg, planet mass--original mass of the Earth
+radius                      = 6.371e6   # m, planet surface radius
+
+zenith_angle                = 48.19     # Cronin+14 Earth-like case with clear sky [arccos(2/3)]
+asf_scalefactor             = 0.375     # absorbed solar flux scale factor = 3/8
+albedo_s                    = 0.0 		# surface albedo
+albedo_pl                   = 0.2   	# Bond albedo (scattering)
+
+P_top                       = 1.0e-5    # bar, Pressure at TOA
+
+##### PROTEUS settings
+
+# Output subdirectory name (relative to output folder)
+dir_output                  = jgr_grid
+
+# Choose times
+time_star                   = 100.0e6  	# yr, time since star formation
+time_planet                 = 0.        # yr, time since MO start
+time_target                 = 4.567e+9  # yr, target time for MO evolution
+
+# SOCRATES spectral file to use (relative to AEOLUS folder)
+spectral_file               = AEOLUS/spectral_files/Mallard/Mallard
+
+# Stellar heating toggle, 0: disabled | 1: enabled
+stellar_heating             = 1
+
+# Time-stepping stuff
+plot_iterfreq               = 15     # Plotting frequency, 0: wait until completion | n: every n iterations
+sspec_dt_update             = 1e5    # Time intervals at which to re-calculate the stellar spectrum
+sinst_dt_update             = 1e2    # Time intervals at which to re-calculate the instellation
+dt_maximum                  = 2e5    # Maximum time-step
+dt_minimum                  = 5e2    # Minimum time-step
+dt_crit                     = 2e4    # Maximum time-step when |F_atm| < F_crit
+dt_method                   = 0      # Time-stepping method, 0: proportional | 1: adaptive | 2: maximum
+dt_propconst                = 40.0   # Proportionality constant for dt_method=0
+
+# Flux convergence scheme and tolerances for surface equilibration
+flux_convergence            = 0         # 0: off | 1: on | 2: restart
+F_atm_bc                    = 0         # Boundary condition choice for F_atm, 0: TOA | 1: Surface
+F_crit                      = 2.1       # Critical flux, below which attempt to stabilise the evolution (0: disabled)
+F_eps                       = 1.0       # W m-2
+F_diff                      = 0.1       # flux fraction
+RF_crit                     = 0.01      # depth fraction
+dTs_atm                     = 30        # K
+skin_d                      = 0.01      # m
+skin_k                      = 2.0       # W m-1 K-1
+prevent_warming             = 1         # Require that the planet only cool down over time, 0: disabled | 1: enabled
+
+# Break at solidification?
+solid_stop                  = 1
+phi_crit                    = 0.005     # melt fraction
+
+# Break at steady state?
+steady_stop                 = 1
+steady_flux                 = 0.7        # Maximum absolute value of F_atm allowed for convergence
+steady_dprel                = 1.0e-10    # Percentage change in melt fraction over time (dp/p)/dt*100
+
+# Method for solving for T(p) profile
+atmosphere_model            = 0         # Atmosphere model to be used, 0: AEOLUS | 1: AGNI
+atmosphere_solve_energy     = 0         # Enable time-stepped atmosphere solution, 0: disabled | 1: enabled 
+atmosphere_surf_state       = 2         # Atmosphere bottom edge boundary condition, 0: free | 1: fixed at T_surf | 2: conductive skin
+
+# Number of levels
+atmosphere_nlev             = 110
+
+# Temperature limits throughout atmosphere [K]
+min_temperature             = 0.5   
+max_temperature             = 5000.0  
+
+# Tropopause type, 0: none | 1: skin temperature | 2: dynamic
+tropopause                  = 1
+
+# Rayleigh scattering, 0: Disabled | 1: Enabled
+insert_rscatter             = 1  
+
+# Approach to atmospheric chemistry, 0: none | 1: offline | 2: online
+atmosphere_chem_type        = 1  
+
+##### SPIDER-specific settings
+
+# N2 partition coefficients, 0: oxidized | 1: reduced
+N2_partitioning             = 0
+
+# Restart setting: 1: Fresh start | 2: Restart from file
+IC_INTERIOR                 = 1
+
+# Gravitational separation of solid and melt phase, 0: off | 1: on
+SEPARATION                  = 1
+
+# Mixing length parameterization: 1: variable | 2: constant
+mixing_length               = 2
+
+# Ultra-thin thermal boundary layer at top, 0: off | 1: on
+PARAM_UTBL                  = 0               # on/off flag
+param_utbl_const            = 1.0E-8          # value if on
+
+# Standard tolerance for solvers
+solver_tolerance            = 1.0e-10
+
+# Maximum absolute surface temperature change [K]
+tsurf_poststep_change       = 20.0  
+
+# Maximum fractional surface temperature change [fraction]
+tsurf_poststep_change_frac  = 0.04  
+
+# Fractional core radius
+planet_coresize             = 0.55
+
+# Restart filename if restart manually, int(yr)+".json"
+ic_interior_filename        = 0
+
+# Initial entropy conditions for interior
+ic_adiabat_entropy          = 2980.0
+ic_dsdr                     = -4.698e-06
+
+# Atmosphere heat flux start guess, adjusted during runtime [W/m^2]
+F_atm                       = 4.0E5
+
+# Oxygen fugacity offset relative to the IW buffer (log10 units)
+fO2_shift_IW                = 0
+
+# Enable solving for initial partial pressures (0: off | 1: on)
+solvepp_enabled             = 1
+
+# Parameters used to solve for initial partial pressures (when solvepp_enabled = 1)
+T_surf_guess                = 3200      # Surface temperature initial guess [K]
+melt_fraction_guess         = 1.0       # Mantle melt fraction initial guess
+CH_ratio                    = 1.0      # C/H ratio
+hydrogen_earth_oceans       = 10.0      # Hydrogen inventory in units of equivalent Earth oceans
+nitrogen_ppmw               = 2.0       # Nitrogen inventory in ppmw relative to mantle mass
+
+# Prescribed injected partial pressures [bar]
+# Summed with solvepp results when solvepp_enabled = 1
+H2O_included                = 1
+H2O_add_bars                = 0.0
+
+CO2_included                = 1
+CO2_add_bars                = 0.0
+
+H2_included                 = 1
+H2_add_bars                 = 0.0
+
+CH4_included                = 1
+CH4_add_bars                = 0.0
+
+CO_included                 = 1
+CO_add_bars                 = 0.0
+
+N2_included                 = 1
+N2_add_bars                 = 0.0
+
+O2_included                 = 0
+O2_add_bars                 = 0.0
+
+S_included                  = 0
+S_add_bars                  = 0.0
+
+He_included                 = 0
+He_add_bars                 = 0.0
+
+# Fractional species melt phase change event trigger [wt]
+H2O_poststep_change         = 0.05  
+CO2_poststep_change         = 0.05
+H2_poststep_change          = 0.05
+CH4_poststep_change         = 0.05
+CO_poststep_change          = 0.05 
+N2_poststep_change          = 0.05
+O2_poststep_change          = 0.05
+S_poststep_change           = 0.05
+He_poststep_change          = 0.05 
+
