+++ proteus.py
@@ -285,10 +285,11 @@ def main():
 
             # Generate a new SOCRATES spectral file containing this new spectrum
             star_spec_src = dirs["output"]+"socrates_star.txt"
-             #    Update stdout
+            #    Update stdout
             old_stdout , old_stderr = sys.stdout , sys.stderr
             sys.stdout = StreamToLogger(log, logging.INFO)
             sys.stderr = StreamToLogger(log, logging.ERROR)
+            #    Spectral file stuff
             PrepareStellarSpectrum(StellarFlux_wl,fl,star_spec_src)
             InsertStellarSpectrum(
                                     spectral_file_nostar,
@@ -296,6 +297,7 @@ def main():
                                     dirs["output"]+"runtime_spectral_file"
                                 )
             os.remove(star_spec_src)
+            #    Restore stdout
             sys.stdout , sys.stderr = old_stdout , old_stderr
 
         else:
