+++ tools/RunPROTEUS.sh
@@ -27,7 +27,6 @@ else
     CFGFILE="$1"
     ALIAS="$2"
     DETACH=$(echo "$3" | tr -d ' ' | tr '[:upper:]' '[:lower:]' | cut -c1-1)  # strip spaces, covert to lowercase, get first char
-    LOGFILE="$COUPLER_DIR/output/$ALIAS.log"
     EXECUTABLE="$COUPLER_DIR/proteus.py"
 
     # Clear dead screens
@@ -41,38 +40,18 @@ else
         exit 1
     fi
 
-    # Setup log file
+    # Setup paths 
     echo "    Config path   = '$CFGFILE' "
-    echo "    Log file path = '$LOGFILE' "
     echo "    Screen alias  = '$ALIAS' "
-
-    rm -f $LOGFILE
-    touch $LOGFILE
 
     # Dispatch screen session with PROTEUS inside
     echo "    Dispatching screen session..."
     COMMAND="python $EXECUTABLE --cfg_file $CFGFILE"
 
-    if [[ "$OSTYPE" == "darwin"* ]]; then
-        # MacOS does not support the -Logfile flag
-        config="log on
-        logfile $LOGFILE";
-        config_file="/tmp/$ALIAS.screenconf"
-        rm -f "$config_file"
-        echo "$config" > "$config_file"
-        if [[ "$DETACH" == "y" ]]; then 
-            screen -S $ALIAS -L -d -m -c "$config_file" bash -c "$COMMAND" 
-        else 
-            screen -S $ALIAS -L -c "$config_file" bash -c "$COMMAND" 
-        fi
-        
-    else
-        # Linux
-        if [[ "$DETACH" == "y" ]]; then 
-            screen -S $ALIAS -d -m -L -Logfile $LOGFILE bash -c "$COMMAND" 
-        else 
-            screen -S $ALIAS -L -Logfile $LOGFILE bash -c "$COMMAND" 
-        fi
+    if [[ "$DETACH" == "y" ]]; then 
+        screen -S $ALIAS -d -m bash -c "$COMMAND" 
+    else 
+        screen -S $ALIAS -c bash -c "$COMMAND" 
     fi
     
     # Done?
